+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Add `s.dict_tag` as a convenience factory for building mapping specs for which
+  the value spec tags are derived automatically from the corresponding dict keys (#52)
 
 ## [v0.2.5] - 2020-04-10
 ### Added